requirements.txt
beautifulsoup4
Flask
requests
gunicorn
Flask-PyMongo
python-dotenv
pydantic